Create a small Artisan command to list calls on a Twilio account
This command isn't that flash. Rather it just connects to a Twilio
account and retrieves a list of calls, then prints them in a table.
There is an argument that allows for the output to be changed; setting
whether the date format should be short or long. Otherwise, it's not
particularly that sophisticated.